Move runtime-safe helpers to shared/ layer (review #5)
Provider src/ no longer imports canonical logic from scripts/lib;
shared/ is the neutral source consumed by src/ and maintainer scripts.
