+++ scripts/lib/cursor-setting-sources.d.mts
@@ -1,5 +1,5 @@
-export declare const CURSOR_SETTING_SOURCES_ENV: "PI_CURSOR_SETTING_SOURCES";
-
-export declare function resolveCursorSettingSources(raw?: string): string[] | undefined;
-
-export declare function serializeCursorSettingSources(settingSources: string[] | undefined): string;
+export {
+	CURSOR_SETTING_SOURCES_ENV,
+	resolveCursorSettingSources,
+	serializeCursorSettingSources,
+} from "../../shared/cursor-setting-sources.mjs";
