+++ scripts/lib/cursor-setting-sources.mjs
@@ -1,22 +1,6 @@
-/** Canonical Cursor settingSources parsing (parity-tested; re-exported from src/cursor-setting-sources.ts). */
-export const CURSOR_SETTING_SOURCES_ENV = "PI_CURSOR_SETTING_SOURCES";
-
-export function resolveCursorSettingSources(raw) {
-	const trimmed = raw?.trim();
-	if (!trimmed) return ["all"];
-	const normalized = trimmed.toLowerCase();
-	if (["0", "false", "off", "none", "omit", "disabled"].includes(normalized)) return undefined;
-	if (["1", "true", "on", "all"].includes(normalized)) return ["all"];
-	const sources = trimmed
-		.split(",")
-		.map((entry) => entry.trim())
-		.filter(Boolean);
-	if (sources.length === 0) return undefined;
-	return sources;
-}
-
-/** Serialize parsed settingSources for PI_CURSOR_SETTING_SOURCES (undefined => explicit none). */
-export function serializeCursorSettingSources(settingSources) {
-	if (settingSources === undefined || settingSources.length === 0) return "none";
-	return settingSources.join(",");
-}
+/** Maintainer-script re-export; canonical implementation lives in shared/cursor-setting-sources.mjs. */
+export {
+	CURSOR_SETTING_SOURCES_ENV,
+	resolveCursorSettingSources,
+	serializeCursorSettingSources,
+} from "../../shared/cursor-setting-sources.mjs";
