+++ shared/cursor-sensitive-text.d.mts
@@ -1,0 +1,1 @@
+export declare function scrubSensitiveText(text: string, apiKey?: string): string;
